refactor(layout): extract providers, inline AppContentLayout, drop CSS modules
Move AppLayoutContext and WikiItemActionsProvider into src/providers/.
Inline AppContentLayout into app/layout.tsx as AppContent (eliminating the
indirection). Delete AppContentLayout.tsx/.module.css and AppLayoutContext.tsx.
Inline WikiItemActions.module.css as Tailwind utilities and delete the file.
Change WikiItemMenu and StarredList to default exports.
