+++ frontend/src/app/app/layout.tsx
@@ -3,12 +3,53 @@
 import { useState, type ReactNode } from "react";
 import { RootLayout } from "@onyx-ai/opal/layouts";
 import AppSidebar from "@/sections/sidebar/AppSidebar";
-import { AppLayoutProvider } from "@/sections/app/AppLayoutContext";
-import { AppContentLayout } from "@/sections/app/AppContentLayout";
+import { AppLayoutProvider, useAppLayout } from "@/providers/AppLayoutProvider";
+import { WikiItemActionsProvider } from "@/providers/WikiItemActionsProvider";
+import { StatusBanner } from "@/sections/app/StatusBanner";
 
 const COLLAPSED_KEY = "agent-wiki:sidebar-collapsed";
 
-export default function AppLayout({ children }: { children: ReactNode }) {
+interface AppContentProps {
+  children: ReactNode;
+}
+
+function AppContent({ children }: AppContentProps) {
+  const {
+    headerContent,
+    leftPanelContent,
+    actionSidebarContent,
+    isActionSidebarOpen,
+  } = useAppLayout();
+  return (
+    <WikiItemActionsProvider>
+      {leftPanelContent && (
+        <RootLayout.LeftPanel>{leftPanelContent}</RootLayout.LeftPanel>
+      )}
+      <RootLayout.App>
+        <StatusBanner />
+        <RootLayout.Header>
+          <div className="flex h-14 items-center px-4">{headerContent}</div>
+        </RootLayout.Header>
+        <RootLayout.MainContent>
+          <div className="mx-auto w-full max-w-(--breakpoint-content-md)">
+            {children}
+          </div>
+        </RootLayout.MainContent>
+      </RootLayout.App>
+      {isActionSidebarOpen && actionSidebarContent && (
+        <RootLayout.RightPanel className="w-60 border-l border-border-01">
+          {actionSidebarContent}
+        </RootLayout.RightPanel>
+      )}
+    </WikiItemActionsProvider>
+  );
+}
+
+interface LayoutProps {
+  children: ReactNode;
+}
+
+export default function Layout({ children }: LayoutProps) {
   const [folded, setFolded] = useState<boolean>(() => {
     if (typeof window === "undefined") return false;
     const stored = window.localStorage.getItem(COLLAPSED_KEY);
@@ -31,7 +72,7 @@ export default function AppLayout({ children }: { children: ReactNode }) {
         <AppSidebar folded={folded} onFoldToggle={toggle} />
       </RootLayout.Sidebar>
       <AppLayoutProvider>
-        <AppContentLayout>{children}</AppContentLayout>
+        <AppContent>{children}</AppContent>
       </AppLayoutProvider>
     </RootLayout.Root>
   );
