+++ frontend/src/components/RequireAdmin.tsx
@@ -24,7 +24,6 @@
 
 import { useEffect, type ReactNode } from "react";
 import { useRouter } from "next/navigation";
-
 import { useRequireAuth } from "@/lib/auth";
 
 export function RequireAdmin({ children }: { children: ReactNode }) {
