+++ frontend/src/components/wiki/WikiHome.tsx
@@ -5,7 +5,6 @@ import { useState, type FormEvent } from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import useSWR from "swr";
-
 import {
   Button,
   Divider,
@@ -23,13 +22,12 @@ import {
   SvgPlusCircle,
 } from "@onyx-ai/opal/icons";
 import { SvgOnyxLogo } from "@onyx-ai/opal/logos";
-
 import { LoadingSpinner } from "@/components/common/LoadingSpinner";
 import { relativeTime } from "@/lib/time";
 import { listTemplateSummaries } from "@/lib/templates";
 import type { DocumentTemplateSummary } from "@/lib/templates";
 import { AI_DRAFT_KEY, generateDraft, type RecentPage } from "@/lib/wiki";
-import { WikiItemMenu } from "./WikiItemActions";
+import WikiItemMenu from "./WikiItemActions";
 import styles from "./WikiHome.module.css";
 
 export function WikiHome() {
