--- frontend/src/sections/app/AppContentLayout.module.css
@@ -1,17 +1,0 @@
-.headerBar {
-  height: 56px;
-  display: flex;
-  align-items: center;
-  padding: 0 16px;
-}
-
-.mainInner {
-  max-width: var(--breakpoint-content-md);
-  margin: 0 auto;
-  width: 100%;
-}
-
-.rightPanel {
-  width: 240px;
-  border-left: 1px solid var(--border-01);
-}
